s3-read-only-policy.json
This file contains an IAM Policy for defining permissions within AWS environments. It specifies which actions are allowed or denied on specific AWS resources for a user, group, or role. The policy uses JSON format to set permissions based on actions, resources, and conditions, ensuring proper access control in line with your organization's security requirements.