trust-policy.json
This file contains an IAM Trust Policy that defines the conditions under which an IAM role can be assumed by another entity, such as another AWS account, a service, or an external identity provider. It establishes a trusted relationship between the role and the entity, specifying the permissions granted for assuming the role.